Revert "Downloadable logs (#152)"
This reverts commit ea4ab3602af2eaa4b5b15b1d5e23bfd6c0d91d57.
