+++ src/lib/analytics.js
@@ -4,4 +4,3 @@ const GoogleAnalytics = {
 };
 
 export default GoogleAnalytics;
-
